# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Product Quotation API",
    "description": "The Product Quotation API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "BaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "number"
          },
          "updatedBy": {
            "type": "number"
          }
        },
        "required": ["id", "createdAt", "updatedAt"]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "code": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string"
          },
          "parentId": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": ["name", "code"]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "code": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string"
          },
          "parentId": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "Category": {
        "allOf": [
          {"$ref": "#/components/schemas/BaseEntity"},
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "parentId": {
                "type": "number"
              },
              "isActive": {
                "type": "boolean",
                "default": true
              },
              "parent": {
                "$ref": "#/components/schemas/Category"
              },
              "children": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            },
            "required": ["name", "code", "isActive"]
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "summary": "Create a new category",
        "tags": ["categories"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "summary": "Get all categories",
        "tags": ["categories"],
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all categories.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "summary": "Get category by ID",
        "tags": ["categories"],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the category.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found."
          }
        }
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "summary": "Update category",
        "tags": ["categories"],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found."
          }
        }
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "summary": "Delete category",
        "tags": ["categories"],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The category has been successfully deleted."
          },
          "404": {
            "description": "Category not found."
          }
        }
      }
    },
...
  }
}